+++ client/src/App.jsx
@@ -1,0 +1,6 @@
+import "./App.css";
+import MainRoutes from "./routes";
+
+const App = () => MainRoutes();
+
+export default App;