+++ client/src/views/pages/auth/forget/index.jsx
@@ -1,0 +1,3 @@
+const Forget = () => <>Forget</>;
+
+export default Forget;