+++ client/src/views/pages/auth/reset/index.jsx
@@ -1,0 +1,3 @@
+const Reset = () => <>Reset</>;
+
+export default Reset;